Building tests for my modules, as well as refactoring
Building tests is quite hard, but ya know it's just what's needed in order to not go crazy at the end and see that your code is pretty much unuseable. Anyway, currently setting up tests for all of the different phases of the GoFishEngine to make sure that they do what I actually want them to do.

Refactoring consists of tidying up the Player, Card, Engine, and GoFishEngine models to reflect the bugs that I have seen on my tests