Comments and finishing up testing.
Comments on the Engine class source codes. Much needed if anyone were to look at, or if I find that I find that I want to know what's going on with the code faster than I would normally. Testing is almost complete and then I can work on new games as well.